moving ssc code to separate source files
