+++ src/EnergyPlus/CMakeLists.txt
@@ -432,6 +432,8 @@ SET( SRC
   PurchasedAirManager.hh
   PVWatts.cc
   PVWatts.hh
+  PVWattsSSC.cc
+  PVWattsSSC.hh
   RefrigeratedCase.cc
   RefrigeratedCase.hh
   ReportSizingManager.cc
