+++ src/EnergyPlus/PVWatts.hh
@@ -1,4 +1,4 @@
-// EnergyPlus, Copyright (c) 1996-2017, The Board of Trustees of the University of Illinois,
+// EnergyPlus, Copyright (c) 1996-2018, The Board of Trustees of the University of Illinois,
 // The Regents of the University of California, through Lawrence Berkeley National Laboratory
 // (subject to receipt of any required approvals from the U.S. Dept. of Energy), Oak Ridge
 // National Laboratory, managed by UT-Battelle, Alliance for Sustainable Energy, LLC, and other
@@ -59,16 +59,11 @@
 #include <EnergyPlus.hh>
 #include <DataSurfaces.hh>
 #include <InputProcessor.hh>
+#include <PVWattsSSC.hh>
 
 namespace EnergyPlus {
 
 namespace PVWatts {
-
-	const Real64 AOI_MIN(0.5);
-	const Real64 AOI_MAX(89.5);
-	const Real64 DTOR(DataGlobals::DegToRadians);
-	enum RADMODE {DN_DF, DN_GH, GH_DF, POA_R, POA_P};
-	const Real64 SMALL(1e-6);
 
 	enum class ModuleType {
 		STANDARD,
@@ -110,227 +105,6 @@ namespace PVWatts {
 		Real64 ignddiff;
 		int sunup;
 	};
-
-	// From here down is from the SAM Simulation core library
-	
-	class pvwatts_celltemp
-	{
-		/*
-		 This class was converted from the original PVWatts subroutine
-		 to store the previous irradiance and cell temp to make the model
-		 give consistent results with the online PVWatts V1.  Previously
-		 all calculations were done over the course of a day, so the
-		 previous module/sun values were saved, but in the single time stamp
-		 version, these changes were not tracked.  apd 2/24/2012
-
-		 Defines function to calculate cell temperature, changed 8/22/2007 to
-		 work with single time stamp data, also see pvsubs2.c
-
-		 This function was converted from a PVFORM version 3.3 subroutine
-		 c     this routine estimates the array temperature given the poa radiation,
-		 c     ambient temperature, and wind speed.  it uses an advanced cell temp
-		 c     model developed by m fuentes at snla.  if the poa insolation is eq
-		 c     zero then set cell temp = 999.
-		 c
-		 passed variables:
-		 inoct = installed nominal operating cell temperature (deg K)
-		 height = average array height (meters)
-		 poa2 = plane of array irradiances (W/m2)
-		 ws2 = wind speeds (m/s)
-		 ambt2 = ambient temperatures (deg C)
-
-		 c  local variables :
-		 c     absorb = absorbtivity
-		 c     backrt = ratio of actual backside heat xfer to theoretical of rack mount
-		 c     boltz = boltzmann's constant
-		 c     cap = capacitance per unit area of module
-		 c     capo = capacitance per unit area of rack mounted module
-		 c     conair = conductivity of air
-		 c     convrt = ratio of total convective heat xfer coef to topside hxc
-		 c     denair = density of air
-		 c     dtime = time step
-		 c     eigen = product of eigen value and time step
-		 c     emmis = emmisivity
-		 ex = ?
-		 c     grashf = grashoffs number
-		 c     hconv = convective coeff of module (both sides)
-		 c     hforce = forced convective coeff of top side
-		 c     hfree = free convective coeff of top side
-		 c     hgrnd = radiative heat xfer coeff from module to ground
-		 hsky = ?
-		 c     iflagc = flag to check if routine has been executed
-		 c     reynld = reynolds number
-		 c     suun = insolation at start of time step
-		 c     suno = previous hours insolation
-		 c     tamb = ambient temp
-		 c     tave = average of amb and cell temp
-		 c     tgrat = ratio of grnd temp above amb to cell temp above amb
-		 c     tgrnd = temperature of ground
-		 c     tmod = computed cell temp
-		 c     tmodo = cell temp for previous time step
-		 c     tsky = sky temp
-		 c     visair = viscosity of air
-		 c     windmd = wind speed at module height
-		 c     xlen = hydrodynamic length of module              */
-
-	private:
-		int j;
-		double height, inoct;
-		double absorb,backrt,boltz,cap,capo,conair,convrt,denair;
-		double eigen,emmis,grashf,hconv,hforce,hfree,hgrnd,reynld,suun;
-		double suno,tamb,tave,tgrat,tgrnd,tmod,tmodo,tsky,visair,windmd,xlen;
-		double hsky,ex;
-		const double &dtime;
-	public:
-		pvwatts_celltemp( double _inoct, double _height, const double& _dTimeHrs);
-		double operator() ( double poa2, double ws2, double ambt2, double fhconv = 1.0 );
-		void set_last_values( double Tc, double poa );
-	};
-
-	struct poaDecompReq {
-		poaDecompReq() : i(0), dayStart(0), stepSize(1), stepScale('h'), doy(-1) {}
-		size_t i; // Current time index
-		size_t dayStart; // time index corresponding to the start of the current day
-		double stepSize;
-		char stepScale; // indicates whether time steps are hours (h) or minutes (m)
-		double* POA; // Pointer to entire POA array (will have size 8760 if time step is 1 hour)
-		double* inc; // Pointer to angle of incident array (same size as POA)
-		double* tilt; // Pointer to angle of incident array (same size as POA)
-		double* zen; // Pointer to angle of incident array (same size as POA)
-		double* exTer; // Pointer to angle of incident array (same size as POA)
-		double tDew;
-		int doy;
-		double elev;
-	};
-
-	class irrad
-	{
-	private:
-		int year, month, day, hour;
-		double minute, delt;
-
-		double lat, lon, tz;
-		int radmode, skymodel, track;
-		double gh, dn, df, wfpoa, alb;
-		double tilt, sazm, rlim, gcr;
-		bool en_backtrack;
-		double sun[9], angle[5], poa[3], diffc[3];
-		int tms[3];
-		double ghi;
-
-		poaDecompReq* poaAll;
-
-	public:
-
-		irrad();
-		int check();
-
-		// if delt_hr is less than zero, do not interpolate sunrise and sunset hours
-#define IRRADPROC_NO_INTERPOLATE_SUNRISE_SUNSET (-1.0)
-		void set_time( int year, int month, int day, int hour, double minute, double delt_hr );
-		void set_location( double lat, double lon, double tz);
-		//skymodel: 0 is isotropic, 1 is hdkr, 2 is perez
-		void set_sky_model( int skymodel, double albedo );
-		void set_surface( int tracking, double tilt_deg, double azimuth_deg, double rotlim_deg, bool en_backtrack, double gcr );
-		void set_beam_diffuse( double beam, double diffuse );
-		void set_global_beam( double global, double beam );
-		void set_global_diffuse(double global, double diffuse);
-		void set_poa_reference( double poa, poaDecompReq* );
-		void set_poa_pyranometer( double poa, poaDecompReq* );
-
-		int calc();
-
-		void get_sun( double *solazi,
-					 double *solzen,
-					 double *solelv,
-					 double *soldec,
-					 double *sunrise,
-					 double *sunset,
-					 int *sunup,
-					 double *eccfac,
-					 double *tst,
-					 double *hextra );
-		void get_angles( double *aoi,
-						double *surftilt,
-						double *surfazi,
-						double *axisrot,
-						double *btdiff );
-		void get_poa( double *beam, double *skydiff, double *gnddiff,
-					 double *isotrop, double *circum, double *horizon );
-		void get_irrad (double *ghi, double *dni, double *dhi);
-		double get_ghi();
-		double get_sunpos_calc_hour();
-	};
-
-	double shade_fraction_1x( double solazi, double solzen,
-							 double axis_tilt, double axis_azimuth,
-							 double gcr, double rotation );
-
-	void diffuse_reduce(double solzen,
-						double stilt,
-						double Gb_nor,
-						double Gd_poa,
-						double gcr,
-						double phi0, // mask angle (degrees)
-						double alb,
-						double nrows,
-
-						double &reduced_skydiff,
-						double &Fskydiff,
-						double &reduced_gnddiff,
-						double &Fgnddiff );
-
-	double iam( double theta_deg, bool ar_glass ); // incidence angle modifier factor relative to normal incidence
-	//static void get_vertices( double axis_tilt, double axis_azimuth, double gcr, double vertices[3][4][3], double rotation);
-
-	//static double vec_dot(double a[3], double b[3]);
-
-	//static void vec_cross(double a[3], double b[3], double result[3]);
-
-	//static void vec_diff( double a[3], double b[3], double result[3] );
-
-	void solarpos(int year,int month,int day,int hour,double minute,double lat,double lng,double tz,double sunn[9]);
-
-	void incidence(int mode,double tilt,double sazm,double rlim,double zen,double azm, bool en_backtrack, double gcr, double angle[5]);
-
-	void isotropic( double hextra, double dn, double df, double alb, double inc, double tilt, double zen, double poa[3], double diffc[3] /* can be NULL */ );
-
-	void perez( double hextra, double dn,double df,double alb,double inc,double tilt,double zen, double poa[3], double diffc[3] /* can be NULL */ );
-
-	void hdkr( double hextra, double dn, double df, double alb, double inc, double tilt, double zen, double poa[3], double diffc[3] /* can be NULL */ );
-
-	void poaDecomp( double wfPOA, double angle[], double sun[], double alb, poaDecompReq* pA, double &dn, double &df, double &gh, double poa[3], double diffc[3]);
-
-	//static void sun_unit( double sazm, double szen, double sun[3] );
-
-	double iam_nonorm( double theta_deg, bool ar_glass );  // non-normalized cover loss (typically use one above!)
-
-	double ModifiedDISC(const double g[3], const double z[3], double td, double alt, int doy, double &dn);
-
-	void ModifiedDISC(const double kt[3], const double kt1[3], const double g[3], const double z[3], double td, int doy, double &dn);
-
-	//static int julian(int yr,int month,int day);
-
-	//static int day_of_year( int month, int day_of_month );
-
-	double backtrack( double solazi, double solzen,
-					 double axis_tilt, double axis_azimuth,
-					 double rotlim, double gcr, double rotation_ideal);
-
-	double GTI_DIRINT( const double poa[3], const double inc[3], double zen, double tilt, double ext, double alb, int doy, double tDew, double elev, double& dnOut, double& dfOut, double& ghOut, double poaCompOut[3]);
-
-	double Min( double v1, double v2);
-
-	double Max( double v1, double v2);
-
-	double transmittance( double theta1_deg, /* incidence angle of incoming radiation (deg) */
-						 double n_cover,  /* refractive index of cover material, n_glass = 1.586 */
-						 double n_incoming, /* refractive index of incoming material, typically n_air = 1.0 */
-						 double k,        /* proportionality constant assumed to be 4 (1/m) for derivation of Bouguer's law (set to zero to skip bougeur's law */
-						 double l_thick,  /* material thickness (set to zero to skip Bouguer's law */
-						 double *_theta2_deg = 0 ); /* thickness of cover material (m), usually 2 mm for typical module */
-
-	// After this is E+ code again.
 
 	class PVWattsGenerator
 	{
